Get current hash from local file and run dev autodeploy every 20min.
I looked at the logs and dev was being deployed more often
than it should have based on changes in master. This is due
to the request to www-dev.m.o/media/revision.txt sometimes
returning a 500 page. This is seen by the script as different
from the hash on github, and thus it deploys. The problem is that
this error is most likely just the previous deploy finishing up since
deploys now take longer than before. This change both reduces the
frequency of autodeploys to every 20 min and checks the local
revision.txt for what is currently deployed. We might consider
moving to a python-based autodeploy script in future for more
flexibility in checking things on the dev server and github.